refactor: split cli-shared into utils/daemon-client and cli.ts
- Moved daemon communication helpers (loadConfig, callDaemon,
  isDaemonRunning, startDaemonProcess, ensureDaemonRunning,
  handleDaemonCommand) from cli-shared.ts to src/utils/daemon-client.ts
- Moved the refund command from cli-shared.ts into cli.ts with the
  rest of the CLI commands
- cli.ts now imports program directly from commander and daemon
  helpers from utils/daemon-client
- TUI imports are updated to use utils/daemon-client
- Removed src/cli-shared.ts
